Conflicts:
	src/FBSession.h
	src/FBSession.m
move to newest sdk (3.11)

+++ src/FBFrictionlessRequestSettings.m
@@ -6,7 +6,7 @@
  * You may obtain a copy of the License at
  *
  *    http://www.apache.org/licenses/LICENSE-2.0
- 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-#import "Facebook.h"
 #import "FBFrictionlessRequestSettings.h"
+
+#import "Facebook.h"
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 // private interface
@@ -29,7 +30,7 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 
-@implementation FBFrictionlessRequestSettings 
+@implementation FBFrictionlessRequestSettings
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 // public
@@ -59,7 +60,7 @@
                                     andDelegate:self];
     if (request) {
         self.activeRequest = request;
-    }    
+    }
 }
 
 - (NSArray *)recipientIDs {
@@ -69,7 +70,7 @@
 - (void)updateRecipientCacheWithRecipients:(NSArray*)ids {
     // if setting recipients directly, no need to complete pending request
     self.activeRequest = nil;
-    
+
     if (ids == nil) {
         self.allowedRecipients = [[[NSArray alloc] init] autorelease];
     } else {
@@ -106,10 +107,10 @@
             @throw [NSException exceptionWithName:NSInvalidArgumentException
                                            reason:@"items in fbids must be NSString or NSNumber"
                                          userInfo:[NSDictionary dictionaryWithObjectsAndKeys:
-                                                   [fbid class], @"invalid class", 
+                                                   [fbid class], @"invalid class",
                                                    nil]];
         }
-        
+
         // if we miss our cache once, we fail the set
         if (![self isFrictionlessEnabledForRecipient:fbidstr]) {
             return NO;
@@ -122,16 +123,16 @@
     // a little request bookkeeping
     self.activeRequest = nil;
 
-    int items = [[result objectForKey: @"data"] count];
+    NSUInteger items = [[result objectForKey: @"data"] count];
     NSMutableArray* recipients = [[[NSMutableArray alloc] initWithCapacity: items] autorelease];
-        
-    for (int i = 0; i < items; i++) {
-        [recipients addObject: [[[result objectForKey: @"data"] 
-                                 objectAtIndex: i] 
+
+    for (NSUInteger i = 0; i < items; i++) {
+        [recipients addObject: [[[result objectForKey: @"data"]
+                                 objectAtIndex: i]
                                 objectForKey: @"recipient_id"]] ;
     }
-        
-    self.allowedRecipients = recipients;        
+
+    self.allowedRecipients = recipients;
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
@@ -142,11 +143,11 @@
 }
 
 - (void)request:(FBRequest *)request didFailWithError:(NSError *)error {
-    // if the request to load the frictionless recipients fails, proceed without updating 
+    // if the request to load the frictionless recipients fails, proceed without updating
     // the recipients cache; the cache may become invalid due to a failed update or other reasons
     // (e.g. simultaneous use of the same app from multiple devices), in the case of an invalid
     // cache, a request dialog may either appear a moment later than it usually would, or appear
-    // briefly when it should not appear at all; in either case the correct request behavior 
+    // briefly when it should not appear at all; in either case the correct request behavior
     // occurs, and the cache is updated
     self.activeRequest = nil;
 }
@@ -154,7 +155,7 @@
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 // NSObject
 
-- (void)dealloc {    
+- (void)dealloc {
     self.activeRequest = nil;
     self.allowedRecipients = nil;
     [super dealloc];
@@ -162,7 +163,7 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 // private helpers
-// 
+//
 
 @synthesize allowedRecipients = _allowedRecipients;
 @synthesize activeRequest = _activeRequest;
