+++ src/FBGraphObject.m
@@ -6,7 +6,7 @@
  * You may obtain a copy of the License at
  *
  *    http://www.apache.org/licenses/LICENSE-2.0
- 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,22 +15,24 @@
  */
 
 #import "FBGraphObject.h"
-#import "FBOpenGraphObject.h"
-#import "FBOpenGraphActionShareDialogParams.h"
+
 #import <objc/runtime.h>
 
+#import "FBOpenGraphActionShareDialogParams.h"
+#import "FBOpenGraphObject.h"
+
 // Module Summary:
-// this is the module that does the heavy lifting to implement the public-facing 
+// this is the module that does the heavy lifting to implement the public-facing
 // developer-model described in FBGraphObject.h, namely typed protocol
 // accessors over "duck-typed" dictionaries, for interating with Facebook Graph and
 // Open Graph objects and actions
 //
 // Message forwarding is used as described here:
 // https://developer.apple.com/library/mac/#documentation/Cocoa/Conceptual/ObjCRuntimeGuide/Articles/ocrtForwarding.html
-// in order to infer implementations for property getter 
+// in order to infer implementations for property getter
 // and setter selectors, backed by objectForKey: and setObject:forKey: respectively
 //
-// The basic flow is straightforward, though there are a number of details; 
+// The basic flow is straightforward, though there are a number of details;
 // The flow is:
 // * System doesn't recognize a selector
 // * It calls methodSignatureForSelector:
@@ -39,7 +41,7 @@
 // * The system passes the invocation to forwardInvocation:
 // * We swap out selectors and invoke
 //
-// Additional details include, deferred wrapping of objects as they are fetched by callers, 
+// Additional details include, deferred wrapping of objects as they are fetched by callers,
 // implementations for common methods such as respondsToSelector and conformsToProtocol, as
 // suggested in the previously referenced documentation
 
@@ -47,7 +49,7 @@ static NSString *const FBIsGraphObjectKey = @"com.facebook.FBIsGraphObjectKey";
 
 // used internally by the category impl
 typedef enum _SelectorInferredImplType {
-    SelectorInferredImplTypeNone  = 0,    
+    SelectorInferredImplTypeNone  = 0,
     SelectorInferredImplTypeGet = 1,
     SelectorInferredImplTypeSet = 2
 } SelectorInferredImplType;
@@ -82,18 +84,18 @@ typedef enum _SelectorInferredImplType {
 #pragma mark Lifecycle
 
 - (id)initWrappingDictionary:(NSDictionary *)jsonObject {
-    self = [super init];    
+    self = [super init];
     if (self) {
         if ([jsonObject isKindOfClass:[FBGraphObject class]]) {
             // in this case, we prefer to return the original object,
             // rather than allocate a wrapper
-            
+
             // we are about to return this, better make it the caller's
             [jsonObject retain];
-            
+
             // we don't need self after all
             [self release];
-            
+
             // no wrapper needed, returning the object that was provided
             return (FBGraphObject*)jsonObject;
         } else {
@@ -195,14 +197,14 @@ typedef enum _SelectorInferredImplType {
 
 - (BOOL)conformsToProtocol:(Protocol *)protocol {
     return  [super conformsToProtocol:protocol] ||
-    ([FBGraphObject isProtocolImplementationInferable:protocol 
+    ([FBGraphObject isProtocolImplementationInferable:protocol
                            checkFBGraphObjectAdoption:YES]);
 }
 
 // returns the signature for the method that we will actually invoke
 - (NSMethodSignature *)methodSignatureForSelector:(SEL)sel {
     SEL alternateSelector = sel;
-    
+
     // if we should forward, to where?
     switch ([FBGraphObject inferredImplTypeForSelector:sel]) {
         case SelectorInferredImplTypeGet:
@@ -215,7 +217,7 @@ typedef enum _SelectorInferredImplType {
         default:
             break;
     }
-    
+
     return [super methodSignatureForSelector:alternateSelector];
 }
 
@@ -238,18 +240,18 @@ typedef enum _SelectorInferredImplType {
             // remove 'set' and trailing ':', and lowercase the new first character
             [propertyName deleteCharactersInRange:NSMakeRange(0, 3)];                       // "set"
             [propertyName deleteCharactersInRange:NSMakeRange(propertyName.length - 1, 1)]; // ":"
-            
+
             NSString *firstChar = [[propertyName substringWithRange:NSMakeRange(0,1)] lowercaseString];
             [propertyName replaceCharactersInRange:NSMakeRange(0, 1) withString:firstChar];
             // the object argument is already in the right place (2), but we need to set the key argument
             [invocation setArgument:&propertyName atIndex:3];
             // and replace the missing method with setObject:forKey:
             invocation.selector = @selector(setObject:forKey:);
-            [invocation invokeWithTarget:self]; 
+            [invocation invokeWithTarget:self];
             break;
-        } 
+        }
         case SelectorInferredImplTypeNone:
-        default: 
+        default:
             [super forwardInvocation:invocation];
             return;
     }
@@ -293,7 +295,7 @@ typedef enum _SelectorInferredImplType {
 }
 
 - (void)setObject:(id)object forKey:(id)key {
-    return [_jsonObject setObject:object forKey:key];    
+    return [_jsonObject setObject:object forKey:key];
 }
 
 - (void)removeObjectForKey:(id)key {
@@ -316,19 +318,19 @@ typedef enum _SelectorInferredImplType {
     } else if ([originalObject isKindOfClass:[NSArray class]]) {
         result = [[[FBGraphObjectArray alloc] initWrappingArray:originalObject] autorelease];
     }
-    
+
     // return our object
     return result;
 }
 
-// helper method used by the catgory implementation to determine whether a selector should be handled 
+// helper method used by the catgory implementation to determine whether a selector should be handled
 + (SelectorInferredImplType)inferredImplTypeForSelector:(SEL)sel {
     // the overhead in this impl is high relative to the cost of a normal property
-    // accessor; if needed we will optimize by caching results of the following 
+    // accessor; if needed we will optimize by caching results of the following
     // processing, indexed by selector
-    
+
     NSString *selectorName = NSStringFromSelector(sel);
-    int	parameterCount = [[selectorName componentsSeparatedByString:@":"] count]-1;
+    NSUInteger parameterCount = [[selectorName componentsSeparatedByString:@":"] count]-1;
     // we will process a selector as a getter if paramCount == 0
     if (parameterCount == 0) {
         return SelectorInferredImplTypeGet;
@@ -337,26 +339,26 @@ typedef enum _SelectorInferredImplType {
                [selectorName hasPrefix:@"set"] &&       // ... we have the proper prefix
                selectorName.length > 4) {               // ... there are characters other than "set" & ":"
         return SelectorInferredImplTypeSet;
-    } 
-    
+    }
+
     return SelectorInferredImplTypeNone;
 }
 
 + (BOOL)isProtocolImplementationInferable:(Protocol*)protocol checkFBGraphObjectAdoption:(BOOL)checkAdoption {
-    // first handle base protocol questions 
+    // first handle base protocol questions
     if (checkAdoption && !protocol_conformsToProtocol(protocol, @protocol(FBGraphObject))) {
         return NO;
     }
-    
+
     if ([protocol isEqual:@protocol(FBGraphObject)]) {
         return YES; // by definition
     }
 
-    unsigned int count = 0;    
+    unsigned int count = 0;
     struct objc_method_description *methods = nil;
-    
+
     // then confirm that all methods are required
-    methods = protocol_copyMethodDescriptionList(protocol, 
+    methods = protocol_copyMethodDescriptionList(protocol,
                                                  NO,        // optional
                                                  YES,       // instance
                                                  &count);
@@ -364,10 +366,10 @@ typedef enum _SelectorInferredImplType {
         free(methods);
         return NO;
     }
-    
+
     @try {
         // fetch methods of the protocol and confirm that each can be implemented automatically
-        methods = protocol_copyMethodDescriptionList(protocol, 
+        methods = protocol_copyMethodDescriptionList(protocol,
                                                      YES,   // required
                                                      YES,   // instance
                                                      &count);
@@ -380,16 +382,16 @@ typedef enum _SelectorInferredImplType {
     } @finally {
         if (methods) {
             free(methods);
-        }   
+        }
     }
-    
+
     // fetch adopted protocols
     Protocol **adopted = nil;
-    @try { 
+    @try {
         adopted = protocol_copyProtocolList(protocol, &count);
         for (int index = 0; index < count; index++) {
             // here we go again...
-            if (![FBGraphObject isProtocolImplementationInferable:adopted[index] 
+            if (![FBGraphObject isProtocolImplementationInferable:adopted[index]
                                        checkFBGraphObjectAdoption:NO]) {
                 return NO;
             }
@@ -399,7 +401,7 @@ typedef enum _SelectorInferredImplType {
             free(adopted);
         }
     }
-    
+
     // protocol ran the gauntlet
     return YES;
 }
@@ -420,13 +422,13 @@ typedef enum _SelectorInferredImplType {
         if ([jsonArray isKindOfClass:[FBGraphObjectArray class]]) {
             // in this case, we prefer to return the original object,
             // rather than allocate a wrapper
-            
+
             // we are about to return this, better make it the caller's
             [jsonArray retain];
-            
+
             // we don't need self after all
             [self release];
-            
+
             // no wrapper needed, returning the object that was provided
             return (FBGraphObjectArray*)jsonArray;
         } else {
@@ -458,8 +460,8 @@ typedef enum _SelectorInferredImplType {
 }
 
 - (void)graphObjectifyAll {
-    int count = [_jsonArray count];
-    for (int i = 0; i < count; ++i) {
+    NSUInteger count = [_jsonArray count];
+    for (NSUInteger i = 0; i < count; ++i) {
         [self graphObjectifyAtIndex:i];
     }
 }
